requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
qrcode>=7.4.0
reportlab>=4.0.0
python-decouple>=3.8
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
dj-database-url>=2.0.0
